+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_inputs_outputs/external_data.py
@@ -110,7 +110,7 @@ class Database(DictMixin, RestTranslatableMixin):
         return result
 
     def _update_name(self, name: str) -> None:
-        self.name = name
+        self.name = name  # type: ignore[assignment]
 
     @classmethod
     def _from_rest_object(cls, obj: Dict) -> "Database":