+++ sdk/netapp/azure-mgmt-netapp/generated_samples/volumes_latest_backup_status.py
@@ -39,6 +39,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-09-01/Volumes_LatestBackupStatus.json
+# x-ms-original-file: 2025-09-01-preview/Volumes_LatestBackupStatus.json
 if __name__ == "__main__":
     main()